feat: Implement `eval` instruction to call native functions
This commit introduces the `eval` instruction, which allows the virtual machine
to execute function calls. The instruction pops a function and its arguments off
the stack, executes the call, and pushes the return value back onto the stack.

As part of this, native (Zig) functions can now be registered in the VM's global
namespace. The `+` builtin is now automatically registered when a new `Vm` is
initialized.